\name{d.biodiversity}

\docType{data}

\alias{d.biodiversity}

\title{Linkages and Collective Action in Biodiversity Conservation}

\description{
This data set is from Basurto (2013), who analyzes the determinants of the 
emergence and endurance of autonomy among local institutions for biodiversity 
conservation in Costa Rica using fsQCA. 
}

\usage{data(d.biodiversity)}

\format{
This data frame contains 30 rows (cases) and the following 9 columns (variables):

\tabular{rlll}{
[ , 1] \tab \bold{AU}  \tab outcome:   \tab local autonomy                                                 \cr
       \tab            \tab            \tab ("1" always, "0" never)                                        \cr
[ , 2] \tab \bold{EM}  \tab condition: \tab local communal involvement through direct employment           \cr   
       \tab            \tab            \tab ("1" 100 percent, "0" 0 percent)                               \cr
[ , 3] \tab \bold{SP}  \tab condition: \tab local direct spending                                          \cr   
       \tab            \tab            \tab ("1" always, "0" never)                                        \cr
[ , 4] \tab \bold{CO}  \tab condition: \tab co-management with local or regional stakeholders              \cr
       \tab            \tab            \tab ("1" present, "0" absent)                                      \cr
[ , 5] \tab \bold{CI}  \tab condition: \tab degree of influence of national civil service policies         \cr
       \tab            \tab            \tab ("1" 100 percent civil service employees, "0" 0 percent)       \cr
[ , 6] \tab \bold{PO}  \tab condition: \tab national participation in policy-making                        \cr
       \tab            \tab            \tab ("1" perceived programme influence, "0" no perceived influence)\cr
[ , 7] \tab \bold{RE}  \tab condition: \tab research-oriented partnerships                                 \cr
       \tab            \tab            \tab ("1" many, "0" few)                                            \cr
[ , 8] \tab \bold{CN}  \tab condition: \tab conservation-oriented partnerships                             \cr
       \tab            \tab            \tab ("1" many, "0" few)                                            \cr
[ , 9] \tab \bold{DE}  \tab condition: \tab direct support by development organizations                    \cr
       \tab            \tab            \tab ("1" always, "0" never)                                                         
}
}

\source{
Basurto, Xavier. 2013. \dQuote{Linking Multi-Level Governance to Local Common-Pool 
Resource Theory using Fuzzy-Set Qualitative Comparative Analysis: Insights from 
Twenty Years of Biodiversity Conservation in Costa Rica.} 
\emph{Global Environmental Change} 23 (3):573-87.
DOI: \href{http://dx.doi.org/10.1016/j.gloenvcha.2013.02.011}{10.1016/j.gloenvcha.2013.02.011}.
}

\keyword{datasets}
