\name{d.CS}

\docType{data}

\alias{d.CS}

\title{Outcomes of Homeless Social Movement Organizations}

\description{
This data set is from Cress and Snow (2000), who analyze the determinants of 
the outcomes attained by homeless social movement organizations using csQCA. 
}

\usage{data(d.CS)}

\format{
This data frame contains 15 rows (cases) and the following 10 columns (variables):

\tabular{rllll}{
[ , 1]  \tab \bold{VI}  \tab condition: \tab viability          \tab ("1" present, "0" absent)\cr
[ , 2]  \tab \bold{DT}  \tab condition: \tab disruptive tactics \tab ("1" present, "0" absent)\cr
[ , 3]  \tab \bold{SA}  \tab condition: \tab sympathetic allies \tab ("1" present, "0" absent)\cr
[ , 4]  \tab \bold{CS}  \tab condition: \tab city support       \tab ("1" present, "0" absent)\cr
[ , 5]  \tab \bold{DF}  \tab condition: \tab diagnostic frame   \tab ("1" present, "0" absent)\cr
[ , 6]  \tab \bold{PF}  \tab condition: \tab prognostic frame   \tab ("1" present, "0" absent)\cr
[ , 7]  \tab \bold{REP} \tab outcome:   \tab representation     \tab ("1" present, "0" absent)\cr
[ , 8]  \tab \bold{RES} \tab outcome:   \tab resources          \tab ("1" present, "0" absent)\cr
[ , 9]  \tab \bold{RIG} \tab outcome:   \tab rights             \tab ("1" present, "0" absent)\cr
[ , 10] \tab \bold{REL} \tab outcome:   \tab relief             \tab ("1" present, "0" absent)
}
}

\source{
Cress, Daniel M., and David A. Snow. 2000. The Outcomes of Homeless Mobilization: 
The Influence of Organization, Disruption, Political Mediation, and Framing. 
\emph{American Journal of Sociology} 105 (4):1063-104.
}

\keyword{datasets}
