\name{d.education}

\docType{data}

\alias{d.education}

\title{Impact of New Public Management Instruments on PhD Education}

\description{
This data set is from Schneider and Sadowski (2010), who analyze the determinants of PhD placement success for 14 economics departments using csQCA. 
}

\usage{data(d.education)}

\format{
This data frame contains 14 rows (cases) and the following 7 columns (variables):

\tabular{rllll}{
[ , 1] \tab \bold{NPM1} \tab condition: \tab local competition      \tab ("1" used, "0" not used)\cr
[ , 2] \tab \bold{NPM2} \tab condition: \tab national competition   \tab ("1" used, "0" not used)\cr
[ , 3] \tab \bold{NPM3} \tab condition: \tab transparency           \tab ("1" used, "0" not used)\cr
[ , 4] \tab \bold{NPM4} \tab condition: \tab university regulations \tab ("1" used, "0" not used)\cr
[ , 5] \tab \bold{NPM5} \tab condition: \tab target agreements      \tab ("1" used, "0" not used)\cr
[ , 6] \tab \bold{NPM6} \tab condition: \tab state regulations      \tab ("1" used, "0" not used)\cr
[ , 7] \tab \bold{O}    \tab outcome:   \tab placement success      \tab ("1" yes, "0" no)
}
}

\source{
Schneider, Peter, and Dieter Sadowski. 2010. \dQuote{The Impact of New Public 
Management Instruments on PhD Education.} \emph{Higher Education} 59 (5):543-65.
DOI: \href{http://dx.doi.org/10.1007/s10734-009-9264-3}{10.1007/s10734-009-9264-3}.
}

\keyword{datasets}
