\name{d.HMN}

\docType{data}

\alias{d.HMN}

\title{The Emergence of the Social Security State}

\description{
This data set is from Hicks, Misra and Ng (1995), who analyze the emergence of
social security programs in 15 industrializing countries during the period
1880-1930 using csQCA.
}

\usage{data(d.HMN)}

\format{
This data frame contains 30 rows (cases) and the following 6 columns (variables):

\tabular{rllll}{
[ , 1]  \tab \bold{LG}  \tab condition: \tab liberal government         \tab ("1" present, "0" absent)\cr
[ , 2]  \tab \bold{CG}  \tab condition: \tab Catholic government        \tab ("1" present, "0" absent)\cr
[ , 3]  \tab \bold{PS}  \tab condition: \tab patriarchal state          \tab ("1" present, "0" absent)\cr
[ , 4]  \tab \bold{UD}  \tab condition: \tab unitary democracy          \tab ("1" present, "0" absent)\cr
[ , 5]  \tab \bold{WM}  \tab condition: \tab working-class mobilization \tab ("1" present, "0" absent)\cr
[ , 6]  \tab \bold{CO}  \tab outcome:   \tab consolidation              \tab ("1" present, "0" absent)
}
}

\source{
Hicks, Alexander, Joya Misra, and Tang N. Ng. 1995. \dQuote{The Programmatic 
Emergence of the Social Security State.} 
\emph{American Sociological Review} 60 (3):329-49.
}

\keyword{datasets}
