\name{d.Kil}

\docType{data}

\alias{d.Kil}

\title{Explaining Urban Regimes}

\description{
This data set is from Kilburn (2004), who analyzes the influence of city context
on urban regimes across 14 cities in the United States using csQCA.
}

\usage{data(d.Kil)}

\format{
This data frame contains 14 rows (cases) and the following 6 columns (variables):

\tabular{rllll}{
[ , 1]  \tab \bold{MLC} \tab condition: \tab mobility of local capital \tab ("1" high, "0" not high)\cr
[ , 2]  \tab \bold{FRB} \tab condition: \tab fiscal resource base      \tab ("1" large, "0" not large)\cr
[ , 3]  \tab \bold{CP}  \tab condition: \tab civic participation       \tab ("1" high, "0" not high)\cr
[ , 4]  \tab \bold{WSR} \tab condition: \tab ward-style representation \tab ("1" high, "0" not high)\cr
[ , 5]  \tab \bold{CS}  \tab condition: \tab city size                 \tab ("1" large, "0" not large)\cr
[ , 6]  \tab \bold{RT}  \tab outcome:   \tab regime type               \tab ("1" progressive, "0" developmental/caretaker)
}
}

\source{
Kilburn, H. Whitt. 2004. \dQuote{Explaining U.S. Urban Regimes.}
\emph{Urban Affairs Review} 39 (5):633-51.
DOI: \href{http://dx.doi.org/10.1177/1078087403262861}{10.1177/1078087403262861}.
}

\keyword{datasets}
