\name{d.Kro}

\docType{data}

\alias{d.Kro}

\title{Women's Representation in Western Democratic Parliaments}

\description{
This data is from Krook (2010), who analyzes the determinants of high women's 
representation in Western-democratic parliaments using csQCA.
}

\usage{data(d.Kro)}

\format{
This data frame contains 22 rows (cases) and the following 6 columns (variables):

\tabular{rlll}{
[ , 1] \tab \bold{ES}  \tab condition: \tab PR electoral system                              \cr
       \tab            \tab            \tab ("1" yes, "0" no)                                \cr
[ , 2] \tab \bold{QU}  \tab condition: \tab quota for women                                  \cr
       \tab            \tab            \tab ("1" yes, "0" no)                                \cr
[ , 3] \tab \bold{WS}  \tab condition: \tab social-democratic welfare system                 \cr
       \tab            \tab            \tab ("1" yes, "0" no)                                \cr
[ , 4] \tab \bold{WM}  \tab condition: \tab autonomous women's movement                      \cr
       \tab            \tab            \tab ("1" yes, "0" no)                                \cr
[ , 5] \tab \bold{LP}  \tab condition: \tab seats held by left-libertarian parties           \cr
       \tab            \tab            \tab ("1" at least 7 percent, "0" less than 7 percent)\cr
[ , 6] \tab \bold{WNP} \tab outcome:   \tab women in single/lower house of parliament        \cr
       \tab            \tab            \tab ("1" at least 30 percent, "0" less than 30 percent)
}
}

\note{
The row names are the official International Organization for Standardization (ISO) country code elements as 
specified in ISO 3166-1-alpha-2.
}

\source{
Krook, Mona Lena. 2010. Women's Representation in Parliament: A Qualitative 
Comparative Analysis 
\emph{Political Studies} 58 (5):886-908.
}


\keyword{datasets}