\name{d.napoleon}

\docType{data}

\alias{d.napoleon}

\title{Reception of the Code Napoleon in Germany}

\description{
This data set is from Arvind and Stirton (2010), who analyze the reception of the
Code Napoleon in Germany using fsQCA.
}

\usage{data(d.napoleon)}

\format{
This data frame contains 14 rows (cases) and the following 8 columns (variables):

\tabular{rllll}{
[ , 1]  \tab \bold{D} \tab condition: \tab legal system              \tab ("1" heterogenous, "0" homogenous)\cr
[ , 2]  \tab \bold{C} \tab condition: \tab territory                 \tab ("1" ruled by France, "0" ruled by enemy)\cr
[ , 3]  \tab \bold{I} \tab condition: \tab state institutions        \tab ("1" strong, "0" none)\cr
[ , 4]  \tab \bold{F} \tab condition: \tab economy                   \tab ("1" seigneural-feudal, "0" proto-industrial)\cr
[ , 5]  \tab \bold{L} \tab condition: \tab ideology of state ruler   \tab ("1" liberal, "0" conservative)\cr
[ , 6]  \tab \bold{N} \tab condition: \tab nativist tendencies       \tab ("1" yes, "0" no)\cr
[ , 7]  \tab \bold{A} \tab condition: \tab sentiments towards France \tab ("1" very negative, "0" very positive)\cr
[ , 8]  \tab \bold{O} \tab outcome:   \tab adoption of Code Napoleon \tab ("1" yes, "0" no)
}
}

\source{
Arvind, Thiruvallore T., and Lindsay Stirton. 2010. \dQuote{Explaining the 
Reception of the Code Napoleon in Germany: A Fuzzy-Set Qualitative Comparative 
Analysis.} \emph{Legal Studies} 30 (1):1-29.
DOI: \href{http://dx.doi.org/10.1111/j.1748-121X.2009.00150.x}{10.1111/j.1748-121X.2009.00150.x}.
}

\keyword{datasets}
