\name{d.stakeholder}

\docType{data}

\alias{d.stakeholder}

\title{Decoupling in Response to Stakeholder Pressures}

\description{
This data set is from Crilly, Zollo and Hansen (2012), who analyze the determinants 
of firms' responses to institutional pressures using fsQCA. 
}

\usage{data(d.stakeholder)}

\format{
This data frame contains 17 rows (cases) and the following 5 columns (variables):

\tabular{rllll}{
[ , 1] \tab \bold{PA} \tab condition: \tab potential for asymmetry \tab ("1" high, "0" low)\cr
[ , 2] \tab \bold{SC} \tab condition: \tab stakeholder consensus   \tab ("1" high, "0" low)\cr
[ , 3] \tab \bold{OI} \tab condition: \tab organizational interest \tab ("1" high, "0" low)\cr
[ , 4] \tab \bold{MC} \tab condition: \tab managerial consensus    \tab ("1" high, "0" low)\cr
[ , 5] \tab \bold{SA} \tab outcome:   \tab substantive action      \tab ("1" high, "0" low)
}
}

\source{
Crilly, Donal, Maurizio Zollo, and Morten T. Hansen. 2012. \dQuote{Faking It or 
Muddling Through? Understanding Decoupling in Response to Stakeholder Pressures.} 
\emph{Academy of Management Journal} 55 (6):1429-48.
DOI: \href{http://dx.doi.org/10.5465/amj.2010.0697}{10.5465/amj.2010.0697}.
}

\keyword{datasets}
