\name{Ebbinghaus}

\docType{data}

\alias{Ebbinghaus}

\title{Union Growth and Decline in Western Europe 1950-95}

\description{
The \bold{Ebbinghaus} data frame has 13 rows and 10 columns.

During the early post-war period, Western trade union movements grew in membership and achieved
an institutionalized role in industrial relations and politics. However, during the last decades,
many trade unions have seen their membership decline as they came increasingly under pressures
due to the social, economic and political changes. This article reviews the main structural,
cyclical and institutional factors explaining union growth and decline. Concentrating on Western
Europe, the empirical analysis compares cross-national union density data for 13 countries over
the first period (1950-75) and for 16 countries over the second, ''crisis'' period (1975-95)
}

\usage{data(Ebbinghaus)}

\format{
The dataset contains the following columns:
\tabular{ll}{
DYNA \tab whether \cr
ACCESS \tab access to media \cr
INFLU \tab influential allies \cr
ELITE \tab divided elites \cr
SOCIAL \tab social networks \cr
OUT \tab the outcome variable
}
}

\source{\url{http://www.compasss.org}}

\references{
Osa, Maryjane and Corduneanu-Huci, Cristina 2003 \emph{Running Uphill: Political Opportunity
in Non-democracies}, Comparative Sociology 2(4), pp. 605-629(25)
}

\keyword{datasets}
