\name{Emme}

\docType{data}

\alias{Emme}

\title{Job Security Regulations in Western Democracies}

\description{
The data frame \bold{Emme} has 19 rows and 7 columns.

This data is from the study by Emmenegger (2011), which analyzes the determinants of 
high job security regulations in Western democracies using fsQCA. 
}

\usage{data(Emme)}

\format{
The data contains the following sets:

\tabular{rl}{
\bold{S}   \tab Condition: high level of statism \cr
\bold{C}   \tab Condition: high level of non-market coordination \cr
\bold{L}   \tab Condition: high level of labour movement strength \cr
\bold{R}   \tab Condition: high level of Catholicism \cr
\bold{P}   \tab Condition: high level of religious party strength \cr
\bold{V}   \tab Condition: many institutional veto points \cr
\bold{JSR} \tab Outcome  : high level of job security regulations \cr
}
}

\note{
The row names are the official International Organization for Standardization (ISO) country code elements as 
specified in ISO 3166-1-alpha-2.
}

\source{
P. Emmenegger. Job Security Regulations in Western Democracies: A Fuzzy Set Analysis.
\emph{European Journal of Political Research}, 50(3):336-364, 2011.
}


\keyword{datasets}
