\name{Osa}

\docType{data}

\alias{Osa}

\title{Running Uphill: Political Opportunity in Non-democracies}

\description{
The \bold{Osa} data frame has 24 rows and 6 columns.

This data is drawn from a study which analyzes twenty-four cases of occurrence/non-occurrence of 
mobilization in non-democratic states to determine conditions of political opportunity in 
high-risk authoritarian contexts. Political opportunity is sensitive to conditions created by 
divided elites, changes in repression, media access, influential allies, and social networks
}

\usage{data(Osa)}

\format{
The dataset contains the following columns:
\tabular{ll}{
DYNA \tab whether \cr
ACCESS \tab access to media \cr
INFLU \tab influential allies \cr
ELITE \tab divided elites \cr
SOCIAL \tab social networks \cr
OUT \tab the outcome variable
}
}

\source{\url{http://www.compasss.org}}

\references{
Osa, Maryjane and Corduneanu-Huci, Cristina 2003 \emph{Running Uphill: Political Opportunity
in Non-democracies}, Comparative Sociology 2(4), pp. 605-629(25)
}

\keyword{datasets}
