\name{QCA-internal}

\alias{.onAttach}

\alias{asNumeric}
\alias{possibleNumeric}

\alias{trimst}
\alias{trimstars}
\alias{nec}
\alias{suf}
\alias{splitstr}
\alias{getName}
\alias{getBigList}
\alias{splitMainComponents}
\alias{splitBrackets}
\alias{removeSingleStars}
\alias{splitPluses}
\alias{splitStars}
\alias{splitTildas}
\alias{solveBrackets}
\alias{simplifyList}
\alias{negateValues}
\alias{removeDuplicates}
\alias{factor.function}
\alias{getNonChars}
\alias{splitProducts}
\alias{insideBrackets}
\alias{outsideBrackets}
\alias{curlyBrackets}
\alias{roundBrackets}



\alias{getSolution}
\alias{createString}
\alias{prettyString}
\alias{rowDominance}
\alias{sortVector}
\alias{sortMatrix}
\alias{writePrimeimp}


\alias{callAllSol}
\alias{callFindSubsets}
\alias{callRemoveRedundants}
\alias{callSolveChart}
\alias{callSuperSubsetMem}
\alias{callSuperSubset}
\alias{callTruthTable}

\title{Internal Functions}

\description{
  The above functions are internal in the QCA package which are not designed to 
  be called directly by the user.
}

\keyword{internal}