\name{readTosmana}

\alias{readTosmana}

\title{Read Tosmana XML Data File}

\description{
Tosmana can save data in XML format. This function reads XML files and coerces
them to R data frame objects.
}

\usage{
readTosmana(filename)
}

\arguments{
  \item{filename}{A string, the path to the XML file}
}

\note{This function requires the XML package}

\keyword{functions}
