\name{Rokkan}

\docType{data}

\alias{Rokkan}

\title{Divided Working-Class Movements}

\description{
The data frame \bold{Rokkan} has 16 rows and 5 columns.

Abridged from Ragin (1987: 129):\cr
The data was used by Rokkan (1970) in his work on nation building in Western Europe.
Rokkan used a "configurational" approach that bears many similarities to the Boolean
approach presented in this work. His main substantive interest was the growth of mass
democracy and the emergence of different cleavage structures in Western European
polities. One outcome that interested him was the division of some working-class
movements in these countries following the Russian Revolution into internationally
oriented wings and some into nationally oriented wings. He considered the distribution
of this outcome important because of its implication for the future of working-class
mobilization (and cleavage structures in general) in Western Europe.
}

\usage{data(Rokkan)}

\format{
The dataset contains the following sets:

\tabular{rl}{
\bold{C} \tab Condition: National church ("1" yes, "0" no)  \cr
\bold{R} \tab Condition: Significant Roman Catholic population and participation in mass education ("1" yes, "0" no) \cr
\bold{L} \tab Condition: State protection of landed interests ("1" yes, "0" no) \cr
\bold{E} \tab Condition: Early state ("1" yes, "0" no) \cr
\bold{S} \tab Outcome  : Major split in working-class movement provoked by Russian Revolution ("1" yes, "0" no) 
}
}

\note{
The row names are the official International Organization for Standardization (ISO) country code elements as 
specified in ISO 3166-1-alpha-2.
}

\source{
C. C. Ragin. \emph{The Comparative Method: Moving beyond Qualitative and Quantitative Strategies}. 
University of California Press, Berkeley, 1987.
}

\references{
S. Rokkan. \emph{Citizens, Elections, Parties}. McKay, New York, 1970.
}

\keyword{datasets}
