\name{Emme}

\docType{data}

\alias{Emme}

\title{Job Security Regulations in Western Democracies}

\description{
The data frame \bold{Emme} has 19 rows and 7 columns.

This data is from the study by Emmenegger (2011), which analyzes the determinants of 
high job security regulations in Western democracies using fsQCA. 
}

\usage{data(Emme)}

\format{
The data contains the following sets:

\tabular{rrl}{
 numeric \tab \bold{S}   \tab Condition: high level of statism\cr
 numeric \tab \bold{C}   \tab Condition: high level of non-market coordination\cr
 numeric \tab \bold{L}   \tab Condition: high level of labour movement strength\cr
 numeric \tab \bold{R}   \tab Condition: high level of Catholicism\cr
 numeric \tab \bold{P}   \tab Condition: high level of religious party strength\cr
 numeric \tab \bold{V}   \tab Condition: many institutional veto points\cr
 numeric \tab \bold{JSR} \tab Outcome  : high level of job security regulations\cr
}
}

\note{
The row names are the official International Organization for Standardization (ISO) country code elements as 
specified in ISO 3166-1-alpha-2.
}

\source{
P. Emmenegger. Job Security Regulations in Western Democracies: A Fuzzy Set Analysis.
\emph{European Journal of Political Research}, 50(3):336-364, 2011.
}


\keyword{datasets}
