\name{QCA-internal}

\alias{sortVector}
\alias{writePrimeimp}
\alias{writeSolution}


\title{Internal Functions}

\description{
  \code{sortVector}, \code{writePrimeimp} and \code{writeSolution} are internal functions of the 
  QCA package which are not designed to be called directly by the user. All of them are used by 
  the main \code{qmcc} function.
}

\keyword{internal}

