\name{QCA-package}

\alias{QCA-package}

\alias{QCA}

\docType{package}

\title{
Starting with version 2.0, the package QCA acts as an aggregator for the various packages dealing with Qualitative Comparative Analysis. For the moment, it is a gateway to the QCAGUI package, but more related packages are expected to be loaded at runtime.
}

\description{
NOTE: the entire namspace and code has been moved to package \pkg{QCAGUI}.

For details, see:

?QCAGUI
}

\details{
    \tabular{ll}{
    Package: \tab QCA\cr
    Type:    \tab Package\cr
    Version: \tab 2.0\cr
    Date:    \tab 2015-12-03\cr
    License: \tab GPL (>= 2)\cr
    }
}

\author{

\bold{Authors}:\cr
Adrian Dusa\cr
Department of Sociology\cr
University of Bucharest\cr
\email{dusa.adrian@unibuc.ro}

Alrik Thiem\cr
(up to version 1.1-3)\cr
Department of Philosophy\cr
University of Geneva\cr
\email{alrik.thiem@unige.ch}

\bold{Maintainer}:\cr
Adrian Dusa
}


\keyword{package}

