\name{readTosmana}

\alias{readTosmana}

\title{Read Tosmana XML data file}

\description{
Tosmana saves it's data by default in XML format. This function reads the XML file and coerces
it to an R dataframe object.
}

\usage{
readTosmana(filename)
}

\arguments{
  \item{filename}{a string, the path to the XML file}
}

\author{
Adrian Dusa\cr
University of Bucharest, Faculty of Sociology and Social Work\cr
dusa.adrian@unibuc.ro
}

\note{This function requires the XML package}



\keyword{misc}
